Create tags_finder.sh
This script takes a fasta file with short tags and look for them on the sequences of a given fastq file using needle from the EMBOSS tools. It will save the resulting alignment profiles in individual files for each tag. In addition, it will count the number of occurrences, allowing one mismatch on the tag-sequence alignment.